aider: The `message_to_json` method in `text/create_messages.py` has been modified to convert the `Message` object into a dictionary before serializing it to JSON.
# Aider chat conversation:

USER: Traceback (most recent call last):
  File "C:\Users\richa\00.Coding\GitHub\active\eleven_labs\testing.py", line 15, in <module>
    handler.handle_chat(role="user", content="Hi there how are you?")
  File "C:\Users\richa\00.Coding\GitHub\active\eleven_labs\data_handler.py", line 28, in handle_chat
    json_message = self.messages.message_to_json(messages)
  File "C:\Users\richa\00.Coding\GitHub\active\eleven_labs\text\create_messages.py", line 56, in message_to_json
    return json.dumps(message)
  File "C:\Users\richa\anaconda3\envs\eleven\lib\json\__init__.py", line 231, in dumps
    return _default_encoder.encode(obj)
  File "C:\Users\richa\anaconda3\envs\eleven\lib\json\encoder.py", line 199, in encode
    chunks = self.iterencode(o, _one_shot=True)
  File "C:\Users\richa\anaconda3\envs\eleven\lib\json\encoder.py", line 257, in iterencode
    return _iterencode(o, 0)
  File "C:\Users\richa\anaconda3\envs\eleven\lib\json\encoder.py", line 179, in default
    raise TypeError(f'Object of type {o.__class__.__name__} '
TypeError: Object of type Message is not JSON serializable

ASSISTANT: The error message indicates that the `Message` object is not JSON serializable. This is because the `json.dumps()` function does not know how to convert the `Message` object into a JSON format.

To fix this, we need to modify the `message_to_json` method in the `text/create_messages.py` file to convert the `Message` object into a dictionary before passing it to `json.dumps()`.

Here is the proposed change:

```python
text/create_messages.py
<<<<<<< ORIGINAL
    def message_to_json(self, message: dict) -> str:
        return json.dumps(message)
=======
    def message_to_json(self, message: Message) -> str:
        return json.dumps(message.dict())
>>>>>>> UPDATED
```

This change uses the `dict()` method provided by the Pydantic `BaseModel` to convert the `Message` object into a dictionary. The `json.dumps()` function can then serialize this dictionary into a JSON string.

Please confirm if this is the change you want to make.